sys.net: refactor types
