+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/types/index.ts
@@ -1,13 +1,14 @@
-export * from './types';
-export * from './types.peer';
-export * from './types.uri';
+export * from './type.peer';
+export * from './type.peer.connection';
 
-export * from './types.event.peer';
-export * from './types.event.group';
-export * from './types.event.peer.local';
-export * from './types.event.peer.conn';
-export * from './types.event.peer.data';
-export * from './types.event.peer.remote';
+export * from './type.event.peer';
+export * from './type.event.group';
+export * from './type.event.peer.local';
+export * from './type.event.peer.conn';
+export * from './type.event.peer.data';
+export * from './type.event.peer.remote';
 
-export * from './types.strategy';
-export * from './types.group';
+export * from './type.strategy';
+export * from './type.group';
+
+export * from './type.uri';
