+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/types/type.event.group.ts
