+++ code/cell.modules/pkg.sys/net/src/web.PeerNetwork/types/type.peer.ts
@@ -1,0 +1,33 @@
+import { t } from './common';
+
+export type PeerId = string; // An identifier of a network peer.
+export type PeerConnectionId = string;
+export type PeerMediaConstraints = t.PartialDeep<MediaStreamConstraints>;
+export type PeerError = { message: string };
+export type PeerModule = { name: string; version: string };
+
+export type PeerStatus = {
+  id: t.PeerId;
+  isOnline: boolean;
+  createdAt: number;
+  signal: t.PeerSignallingEndpoint;
+  connections: t.PeerConnectionStatus[];
+};
+
+export type PeerSignallingEndpoint = {
+  host: string;
+  port: number;
+  path?: string;
+  secure: boolean;
+  key: string;
+  url: { base: string; peers: string };
+};
+
+/**
+ * Filter on a peer connection.
+ */
+export type PeerFilter = (e: PeerFilterArgs) => boolean;
+export type PeerFilterArgs = {
+  peer: t.PeerId;
+  connection: { id: t.PeerConnectionId; kind: t.PeerConnectionKind };
+};
